shift functions to data_process_funcs.py,importing of former file, separation of features & labels of dataset, modifications of gitignore and stream.py
