+++ .gitignore
@@ -1,2 +1,4 @@
 inputs/*
 log.txt
+__pycache__
+*.pyc
